fix: address PR review comments
Resolves review thread PRRT_kwDORy0Fc86HsTI8

Resolves review thread PRRT_kwDORy0Fc86HsjPT

Resolves review thread PRRT_kwDORy0Fc86HsTI_

Resolves review thread PRRT_kwDORy0Fc86HsTJB

Resolves review thread PRRT_kwDORy0Fc86HsTJE

Resolves review thread PRRT_kwDORy0Fc86HsTJF
